% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{djiad}
\alias{djiad}
\title{Daily data the Dow Jones Industrial Average index}
\format{
An object of class \code{"zoo"} containing the daily closing prices of the DJIA index from January 1985
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
djiad
}
\description{
The daily closing prices of the DJIA index from January 1985
}
\seealso{
Other data sets: 
\code{\link{djiam}},
\code{\link{sp500d}},
\code{\link{sp500m}}
}
\concept{data sets}
\keyword{datasets}
