% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_filtering_alg}
\alias{run_filtering_alg}
\title{Runs the filtering algorithm to identify Bull and Bear states}
\usage{
run_filtering_alg(index)
}
\arguments{
\item{index}{vector containing the stock price index}
}
\value{
A logical vector that contains TRUE for Bull states and FALSE for Bear states
}
\description{
This function implements the filtering algorithm of Lunde and Timmermann (2004) to identify Bull and Bear states
}
\note{
Be aware that the states in the beginning and in the end of \code{"index"} are not properly defined
}
\references{
Lunde, A. and Timmermann, A. (2004). Duration Dependence in Stock Prices: An Analysis
of Bull and Bear Markets. Journal of Business and Economic Statistics, 22 (3), 253-273.
}
