% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setpar_filtering_alg}
\alias{setpar_filtering_alg}
\title{Sets the paramters of the filtering algorithm}
\usage{
setpar_filtering_alg(tr_bull, tr_bear)
}
\arguments{
\item{tr_bull}{threshold to idenitfy a Bull state (in percentages)}

\item{tr_bear}{threshold to idenitfy a Bear state (in percentages)}
}
\value{
None
}
\description{
This function sets the paramters of the filtering algorithm of Lunde and Timmermann (2004)
}
\references{
Lunde, A. and Timmermann, A. (2004). Duration Dependence in Stock Prices: An Analysis
of Bull and Bear Markets. Journal of Business and Economic Statistics, 22 (3), 253-273.
}
