% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sp500d}
\alias{sp500d}
\title{Daily data the S&P 500 index}
\format{
An object of class \code{"zoo"} containing daily closing prices of the S&P 500 index starting from January 1950
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
sp500d
}
\description{
The daily closing prices of the S&P 500 index
}
\seealso{
Other data sets: 
\code{\link{djiad}},
\code{\link{djiam}},
\code{\link{sp500m}}
}
\concept{data sets}
\keyword{datasets}
