% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sp500m}
\alias{sp500m}
\title{Monthly data on the S&P 500 index}
\format{
An object of class \code{"zoo"} containing monthly closing prices of the S&P 500 index starting from January 1950
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
sp500m
}
\description{
Monthly closing prices of the S&P 500 index
}
\seealso{
Other data sets: 
\code{\link{djiad}},
\code{\link{djiam}},
\code{\link{sp500d}}
}
\concept{data sets}
\keyword{datasets}
